Return the first error in json decode
It appeared that because we are adding a value field parser
we drop information about the actual error. Here we introduce
a new implementation of mplus (and thus Alternative) that will
return the first failure.

Unfortunately Parser internals are not exposed, so we had to
introduce a hack to overcome that. After the library will switch
to aeson-2 support and aeson will export Parser the hack can be
removed.

Before the patch:

> eitherDecode @(WithField "id" Int (OnlyField "foo" String)) "{\"id\":1,\"foo\":2}"
> Left "Error in $: key \"value\" not found"

After the patch:

> eitherDecode @(WithField "id" Int (OnlyField "foo" String)) "{\"id\":1,\"foo\":2}"
> Left "Error in $.foo: expected String, but encountered Number"

Fixes #3